+++ tests/transforms/decompositions/test_rz_phase_gradient.py
@@ -20,13 +20,13 @@ import numpy as np
 import pytest
 
 import pennylane as qp
-from pennylane.labs.transforms.decomp_rz_phase_gradient import (
-    make_rz_to_phase_gradient_decomp,
-    validate_phase_gradient_wires,
-)
 from pennylane.ops.functions.assert_valid import _test_decomposition_rule
 from pennylane.tape.plxpr_conversion import CollectOpsandMeas
 from pennylane.transforms.decompose import DecomposeInterpreter
+from pennylane.transforms.decompositions import (
+    make_rz_to_phase_gradient_decomp,
+    validate_phase_gradient_wires,
+)
 from pennylane.wires import WireError
